// src/app/modules/home/modules/order/models/order-location-table.model.ts
export class OrderLocationTableModel {

    id: number;
    order_id: number;
    name: string;
    address: string;
    lanmark: string;
    pincode: string;
    email: string;
    phone: string;
    location_date: Date;
    created_by: string;
    created_date: Date;
    updated_by: string;
    updated_date: Date;
    is_delete: number;
}